backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
supabase
python-dotenv
aiogram
python-multipart
